refactor(console): split detect.rs into detect/ submodules (≤500-line cap, refs #926)
detect.rs (648 lines) replaced by detect/ module:
- detect/mod.rs   (61L)  — re-export facade + all_connectors() + registry test
- detect/helpers.rs (212L) — binary_on_path, read_addr_file, tcp_probe,
                             fetch_health_version, detect_service + unit tests
- detect/search.rs  (152L) — SearchConnector + tests
- detect/memory.rs  (146L) — MemoryConnector + tests
- detect/analyze.rs (201L) — AnalyzeConnector + tests

All 17 tests pass; cargo clippy + check_line_cap.sh clean.
